+++ src/schemathesis/cli/commands/run/hypothesis.py
@@ -1,42 +1,13 @@
 from __future__ import annotations
 
-from enum import Enum, unique
 from typing import TYPE_CHECKING, Any
+
+from schemathesis.config._health_check import HealthCheck
 
 if TYPE_CHECKING:
     import hypothesis
 
 HYPOTHESIS_IN_MEMORY_DATABASE_IDENTIFIER = ":memory:"
-
-# Importing Hypothesis is expensive, hence we re-create the enums we need in CLI commands definitions
-# Hypothesis is stable, hence it should not be a problem and adding new variants should not be automatic
-
-
-@unique
-class HealthCheck(str, Enum):
-    # We remove not relevant checks
-    data_too_large = "data_too_large"
-    filter_too_much = "filter_too_much"
-    too_slow = "too_slow"
-    large_base_example = "large_base_example"
-    all = "all"
-
-    def as_hypothesis(self) -> list[hypothesis.HealthCheck]:
-        from hypothesis import HealthCheck
-
-        if self.name == "all":
-            return list(HealthCheck)
-
-        return [HealthCheck[self.name]]
-
-
-def prepare_health_checks(
-    hypothesis_suppress_health_check: list[HealthCheck] | None,
-) -> list[hypothesis.HealthCheck] | None:
-    if hypothesis_suppress_health_check is None:
-        return None
-
-    return [entry for health_check in hypothesis_suppress_health_check for entry in health_check.as_hypothesis()]
 
 
 def prepare_phases(no_shrink: bool = False) -> list[hypothesis.Phase] | None:
@@ -49,11 +20,12 @@ def prepare_phases(no_shrink: bool = False) -> list[hypothesis.Phase] | None:
 
 
 def prepare_settings(
+    *,
     database: str | None = None,
     derandomize: bool | None = None,
     max_examples: int | None = None,
     phases: list[hypothesis.Phase] | None = None,
-    suppress_health_check: list[hypothesis.HealthCheck] | None = None,
+    suppress_health_check: list[HealthCheck],
 ) -> hypothesis.settings:
     import hypothesis
     from hypothesis.database import DirectoryBasedExampleDatabase, InMemoryExampleDatabase
@@ -64,7 +36,10 @@ def prepare_settings(
             ("derandomize", derandomize),
             ("max_examples", max_examples),
             ("phases", phases),
-            ("suppress_health_check", suppress_health_check),
+            (
+                "suppress_health_check",
+                [check for item in suppress_health_check for check in item.as_hypothesis()],
+            ),
         )
         if value is not None
     }